Fixes BZ #1176423: create reservation for VIP ip addresses in DHCP
https://bugzilla.redhat.com/show_bug.cgi?id=1176423

For VIPs on the provisioning network (with DHCP active), calling unused_ip
is insufficient. We also need to create the actual DHCP reservation
so the allocated IP address won't be reused elsewhere.

This change handles both reserving and freeing IP addresses. For a given
VIP nic, we reserve the IP address in DHCP when initially creating this
VIP on a DHCP network, or moving its network traffic type to such a network.
We delete the reservation when moving the VIP *away from* a DHCP network,
or on deleting it.

One side effect of this is VIP creation takes still longer than before, due
to the extra DHCP proxy action, but we already have a "waiting..." dialog with
spinner for this step.